example/requirements.txt
# Example dependencies for the MultiBaas Python SDK
python-dotenv>=1.0.0